requirements.txt
numpy
pandas
plotly
Flask
google-cloud-storage
gunicorn